requirements.txt
# -------------------------
# Core ML / DL
# -------------------------
torch
torchvision
torchaudio

# -------------------------
# Vision
# -------------------------
ultralytics
opencv-python
Pillow

# -------------------------
# NLP / RAG
# -------------------------
transformers
sentence-transformers
faiss-cpu
scikit-learn

# -------------------------
# Audio
# -------------------------
librosa
soundfile
pydub
openai-whisper
faster-whisper

# -------------------------
# UI / Reporting
# -------------------------
streamlit
reportlab
python-dotenv
markdown

# -------------------------
# LLM
# -------------------------
groq
openai

# -------------------------
# Utilities
# -------------------------
numpy
tqdm
